requirements.txt
# =============================================================================
# DJANGO HACKERZ E-COMMERCE - REQUIREMENTS
# =============================================================================

# -----------------------------------------------------------------------------
# Django Core
# -----------------------------------------------------------------------------
Django==5.0.1
Pillow==10.1.0

# -----------------------------------------------------------------------------
# Forms & UI
# -----------------------------------------------------------------------------
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-jazzmin==2.6.0

# -----------------------------------------------------------------------------
# API REST Framework
# -----------------------------------------------------------------------------
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7

# -----------------------------------------------------------------------------
# Content & Document Processing
# -----------------------------------------------------------------------------
markdown==3.5.1
bleach==6.2.0
xhtml2pdf==0.2.16
reportlab==4.0.9

# -----------------------------------------------------------------------------
# HTTP & External API
# -----------------------------------------------------------------------------
requests==2.31.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
